Node class is added to the Linked List class
