+++ DoublyLinkedList/DoublyLinkedList.js
@@ -1,4 +1,10 @@
-const Node = require("./Node.js");
+class Node {
+  constructor(data, next = null, prev = null) {
+    this.data = data;
+    this.next = next;
+    this.prev = prev;
+  }
+}
 
 class DoublyLinkedList {
   constructor() {
