--- DoublyLinkedList/Node.js
@@ -1,9 +1,0 @@
-class Node {
-  constructor(data, next = null, prev = null) {
-    this.data = data;
-    this.next = next;
-    this.prev = prev;
-  }
-}
-
-module.exports = Node;
