+++ LinearLinkedLists/LinkedList.js
@@ -1,4 +1,9 @@
-const Node = require("./Node.js");
+class Node {
+  constructor(data, next = null) {
+    this.data = data;
+    this.next = next;
+  }
+}
 
 class LinkedList {
   constructor(data = null) {
