--- LinearLinkedLists/Node.js
@@ -1,8 +1,0 @@
-class Node {
-  constructor(data, next = null) {
-    this.data = data;
-    this.next = next;
-  }
-}
-
-module.exports = Node;
